Rename core package to "@casterly/core" and move cli to "casterly"
