+++ packages/casterly/README.md
@@ -1,1 +1,10 @@
-# casterly
+# `casterly`
+
+> TODO: description
+
+## Usage
+
+```
+$ casterly watch
+$ casterly build
+```
