+++ packages/casterly/index.js
@@ -1,7 +1,1 @@
-'use strict'
-
-if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./lib/server/defaultServer')
-} else {
-  module.exports = require('./lib/server/devServer')
-}
+module.exports = require('./lib/src/index')
