--- packages/cli/.eslintrc.json
@@ -1,5 +1,0 @@
-{
-  "parserOptions": {
-    "project": "./packages/cli/tsconfig.json"
-  }
-}
