--- packages/cli/README.md
@@ -1,11 +1,0 @@
-# `@casterly/cli`
-
-> TODO: description
-
-## Usage
-
-```
-const cli = require('@casterly/cli');
-
-// TODO: DEMONSTRATE API
-```
