--- packages/cli/index.d.ts
@@ -1,1 +1,0 @@
-export * from './lib/src'
