--- packages/cli/package.json
@@ -1,105 +1,0 @@
-{
-  "name": "@casterly/cli",
-  "version": "0.5.3",
-  "description": "CLI for Casterly",
-  "author": "Lucas Cordeiro <ecdb.lucas@gmail.com>",
-  "homepage": "https://github.com/lucasecdb/casterly#readme",
-  "license": "MIT",
-  "bin": {
-    "casterly": "./bin/casterly-bin.js"
-  },
-  "directories": {
-    "lib": "lib",
-    "test": "__tests__"
-  },
-  "files": [
-    "lib",
-    "bin",
-    "*.js",
-    "*.d.ts"
-  ],
-  "publishConfig": {
-    "registry": "https://registry.npmjs.org",
-    "access": "public"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/lucasecdb/casterly.git",
-    "directory": "packages/cli"
-  },
-  "scripts": {
-    "test": "echo \"Error: run tests from root\" && exit 1",
-    "clean": "rimraf lib",
-    "ts": "tsc",
-    "build": "yarn clean && yarn ts",
-    "dev": "tsc --watch --preserveWatchOutput",
-    "prepublishOnly": "yarn build"
-  },
-  "bugs": {
-    "url": "https://github.com/lucasecdb/casterly/issues"
-  },
-  "dependencies": {
-    "@babel/core": "^7.11.4",
-    "@babel/plugin-proposal-class-properties": "^7.10.4",
-    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/plugin-transform-destructuring": "^7.10.4",
-    "@babel/plugin-transform-runtime": "^7.11.0",
-    "@babel/preset-env": "^7.11.0",
-    "@babel/preset-react": "^7.12.5",
-    "@babel/preset-typescript": "^7.10.4",
-    "@casterly/utils": "^0.5.3",
-    "@pmmmwh/react-refresh-webpack-plugin": "^0.4.3",
-    "@svgr/webpack": "^5.4.0",
-    "babel-core": "^7.0.0-bridge.0",
-    "babel-loader": "^8.0.6",
-    "babel-plugin-macros": "^3.0.0",
-    "babel-plugin-named-asset-import": "^0.3.7",
-    "bfj": "^7.0.1",
-    "case-sensitive-paths-webpack-plugin": "^2.3.0",
-    "chalk": "^4.1.0",
-    "cors": "^2.8.5",
-    "css-loader": "^5.0.1",
-    "css-minimizer-webpack-plugin": "^1.1.5",
-    "dotenv": "^8.2.0",
-    "dotenv-expand": "^5.1.0",
-    "express": "^4.17.1",
-    "file-loader": "^6.2.0",
-    "filesize": "^6.1.0",
-    "fork-ts-checker-webpack-plugin": "^6.0.5",
-    "fs-extra": "^9.0.1",
-    "gzip-size": "^6.0.0",
-    "ignore-loader": "^0.1.2",
-    "mini-css-extract-plugin": "^1.3.0",
-    "nanoid": "^3.1.16",
-    "node-sass": "^5.0.0",
-    "postcss": "^8.1.6",
-    "postcss-flexbugs-fixes": "^5.0.1",
-    "postcss-loader": "^4.0.4",
-    "postcss-preset-env": "^6.7.0",
-    "postcss-safe-parser": "^5.0.0",
-    "react-refresh": "^0.9.0",
-    "recursive-readdir": "^2.2.2",
-    "resolve": "^1.18.1",
-    "sass-loader": "^10.0.5",
-    "semver": "^7.3.2",
-    "strip-ansi": "^6.0.0",
-    "terser-webpack-plugin": "^5.0.3",
-    "unistore": "^3.5.2",
-    "url-loader": "^4.1.0",
-    "webpack": "^5.10.0",
-    "webpack-hot-middleware": "^2.25.0"
-  },
-  "devDependencies": {
-    "@types/chalk": "^2.2.0",
-    "@types/cors": "^2.8.8",
-    "@types/css-minimizer-webpack-plugin": "^1.1.1",
-    "@types/express": "^4.17.8",
-    "@types/fs-extra": "^9.0.3",
-    "@types/node": "^14.14.6",
-    "@types/recursive-readdir": "^2.2.0",
-    "@types/resolve": "^1.17.1",
-    "@types/semver": "^7.3.4",
-    "@types/webpack-hot-middleware": "^2.25.3",
-    "rimraf": "^3.0.0"
-  }
-}
