+++ packages/core/.eslintrc.json
@@ -1,0 +1,5 @@
+{
+  "parserOptions": {
+    "project": "./packages/core/tsconfig.json"
+  }
+}
