+++ packages/core/README.md
@@ -1,0 +1,3 @@
+# `@casterly/core`
+
+> TODO: description
