+++ packages/core/server/defaultServer.ts
@@ -2,9 +2,9 @@ import * as fs from 'fs'
 import * as path from 'path'
 import { parse as parseUrl } from 'url'
 
-import type { RoutesManifest } from '@casterly/cli'
 import { RootContext } from '@casterly/components'
 import { constants, paths } from '@casterly/utils'
+import type { RoutesManifest } from 'casterly'
 import fresh from 'fresh'
 
 import { MAX_AGE_LONG } from '../utils/maxAge'
