+++ packages/core/server/serveStatic.ts
